+++ lib/sec_id.rb
@@ -5,6 +5,7 @@ require 'sec_id/version'
 require 'sec_id/base'
 require 'sec_id/isin'
 require 'sec_id/cusip'
+require 'sec_id/sedol'
 
 module SecId
   Error = Class.new(StandardError)